% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{c19fh_meta}
\alias{c19fh_meta}
\title{Parse model metadata from COVID-19 Forecast Hub repository}
\usage{
c19fh_meta(team = NULL, write = TRUE)
}
\arguments{
\item{team}{Optional indvidual team name for which the metadata; default \code{NULL}}

\item{write}{Boolean indicating whether or not the metadata results should be written to disk}
}
\value{
A \code{tibble} with the model metadata, which includes (among other fields) team name, model name, description of methods, licensing info, and relevant URLs.
}
\description{
Parse model metadata from COVID-19 Forecast Hub repository
}
