% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_arima_params}
\alias{extract_arima_params}
\title{Extract ARIMA parameters}
\usage{
extract_arima_params(arimafit)
}
\arguments{
\item{arimafit}{A single-row mable (\code{mdl_df}) from \code{fabeltools::model(arima=ARIMA(...))}.}
}
\value{
A single-row \code{tibble} containing ARIMA model parameter and diagnostic information.
}
\description{
Extracts ARIMA model parameters, including p, d, q, P, D, Q, and results from \link[broom]{tidy} and \link[broom]{glance} on an ARIMA model object.
}
