% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission.R
\name{format_fit_for_submission}
\alias{format_fit_for_submission}
\title{Title}
\usage{
format_fit_for_submission(.forecast, target_name)
}
\arguments{
\item{target_name}{}
}
\description{
Title
}
\examples{

}
