% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_cases}
\alias{get_cases}
\title{Retrieve case count data}
\usage{
get_cases(source = "jhu", granularity = "national")
}
\arguments{
\item{source}{Data source to query; must be one of \code{'jhu'} or \code{'nyt'}; default is \code{'jhu'}}

\item{granularity}{Data aggregation level; must be one of \code{'national'}, \code{'state'}, or \code{'county'}; if data source is \code{'nyt'} then only \code{'national'} can be used currently; default is \code{'national'}}
}
\value{
A \code{tibble}
}
\description{
Retrieve case count data
}
\references{
https://github.com/CSSEGISandData/COVID-19

https://github.com/nytimes/covid-19-data
}
