% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_hosp}
\alias{get_hosp}
\title{Retrieve hospitalization data}
\usage{
get_hosp(source = "covidtracking", granularity = "national")
}
\arguments{
\item{source}{Data source to query (currently only "covidtracking" is accepted).}

\item{granularity}{Data aggregation level; must be "national".}
}
\value{
a tibble with the following columns:
\itemize{
\item \strong{location}: Currently, "US" only.
\item \strong{epiyear}: Epidemiological year (see \link[lubridate]{epiyear} for more details)
\item \strong{epiweek}: Epidemiological week (see \link[lubridate]{epiweek} for more details)
\item \strong{ihosp}: That week's incident hospitalization increase
}
}
\description{
Retrieve hospitalization data
}
