% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_vax}
\alias{get_vax}
\title{Retrieve vaccination data}
\usage{
get_vax(source = "owid")
}
\arguments{
\item{source}{Data source (currently only "owid" is accepted).}
}
\value{
a tibble with the following columns:
\itemize{
\item \code{location}: "US", State-level FIPS codes, and a few others currently without FIPS codes (e.g., "Indian Health Svc", "Bureau of Prisons", "Dept of Defense", others).
\item \code{epiyear}: Epidemiological year (see \link[lubridate]{epiyear} for more details)
\item \code{epiweek}: Epidemiological week (see \link[lubridate]{epiweek} for more details)
\item \code{total_distributed}: the total number of vaccines distributed to this location
\item \code{total_vaccinations}: the number of total shots given
\item \code{proportion_vaccinations_given}: Ratio, \code{total_vaccinations/total_distributed}
\item \code{people_vaccinated}: the number of people given at least one shot
\item \code{people_fully_vaccinated}: the number of people given the full course of shot+booster
\item \code{population}: the population size of this location
\item \code{total_vaccinations_percap}: the total number of vaccinations given per capita
}
}
\description{
Retrieves vaccination data from \href{https://github.com/owid/covid-19-data}{Our World in Data} (CC-BY).
}
\examples{
v <- get_vax()
v
\dontrun{
library(dplyr)
library(tidyr)
library(ggplot2)
v \%>\%
  filter(location \%in\% c("US", "48", "51")) \%>\%
  mutate(monday=MMWRweek::MMWRweek2Date(epiyear, epiweek, 2)) \%>\%
  group_by(location) \%>\%
  top_n(1, wt=monday) \%>\%
  gather(key, value, total_distributed:total_vaccinations_percap) \%>\%
  ggplot(aes(location, value)) +
  geom_col() +
  facet_wrap(~key, scales="free_y") +
  scale_y_continuous(labels=scales::number_format()) +
  theme_gray()
  }
}
