% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_tsibble}
\alias{make_tsibble}
\title{Make tsibble}
\usage{
make_tsibble(df, chop = TRUE)
}
\arguments{
\item{df}{A tibble containing columns \code{epiyear} and \code{epiweek}.}

\item{chop}{Logical indicating whether or not to remove the most current week (default \code{TRUE}).}
}
\value{
A tsibble containing additional colums \code{monday} indicating the date
for the Monday of that epiweek, and \code{yweek} (a yearweek vctr class object)
that indexes the tsibble in 1 week increments.
}
\description{
Make tsibble
}
