% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quibble}
\alias{quibble}
\title{Make quantile tibbles}
\usage{
quibble(x, q = c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99))
}
\arguments{
\item{x}{A numeric vector.}

\item{q}{Quantiles to report. Defaults to \code{c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99)}.}
}
\value{
A tibble of quantiles (\code{q}) and their values (\code{x}).
}
\description{
Make a quantile tibble for required quantiles.
Defaults to \code{c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99)}.
For \code{N wk ahead inc case target}, filter to: \code{c(0.025, 0.100, 0.250, 0.500, 0.750, 0.900, 0.975)}
}
\examples{
quibble(iris$Sepal.Length)
}
