% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve}
\alias{retrieve}
\alias{get_cases}
\alias{get_deaths}
\title{Retrieve data}
\usage{
get_cases(source = "jhu", granularity = "national")

get_deaths(source = "jhu", granularity = "national")
}
\arguments{
\item{source}{Data source to query; must be one of \code{'jhu'} or \code{'nyt'}; default is \code{'jhu'}}

\item{granularity}{Data aggregation level; must be one of \code{'national'}, \code{'state'}, or \code{'county'}; if data source is \code{'nyt'} then only \code{'national'} can be used currently; default is \code{'national'}}
}
\value{
A \code{tibble} with (at minimum) the following columns:
\itemize{
\item \strong{epiyear}: Epidemiological year (see \link[lubridate]{epiyear} for more details)
\item \strong{epiweek}: Epidemiological week (see \link[lubridate]{epiweek} for more details)
\item \strong{icases/ideaths}: Incident counts (cases or deaths)
\item \strong{ccases/cdeaths}: Cumulative counts (cases or deaths)
}

If \code{source = 'jhu'} and \code{granularity = 'state'} then the \strong{location} column  will include the full name of the state. If \code{source = 'jhu'} and \code{granularity = 'county'} then the \strong{location} column  will include fips (county code).
}
\description{
The package includes functions to retrieve observed data from two canonical sources: the New York Times and the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University. Both organizations administer data aggregation efforts that post daily COVID-19 case and death to GitHub. The data retrieval functions allow the user to specify "source" and "granularity" of data. Internally each function builds a path to the appropriate \code{.csv} file on GitHub, then reads the data into memory. The returned object is data aggregated weekly (using \link[lubridate]{epiweek} and \link[lubridate]{epiyear} designations) for available locations at the granularity specified (national, state, or county level).
}
\references{
https://github.com/CSSEGISandData/COVID-19

https://github.com/nytimes/covid-19-data
}
