% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spread_value}
\alias{spread_value}
\title{Reshape data for submission summary}
\usage{
spread_value(.data, ...)
}
\arguments{
\item{.data}{Tibble with submission data}

\item{...}{Additional arguments passed to \link[tidyr]{spread}}
}
\value{
A \code{tibble} with wide summary data.
}
\description{
This unexported helper function is used in \link[focustools]{submission_summary}. It spreads forecast targets to a wide format and forces "US" locations to be at the top of the resulting \code{tibble}.
}
