% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{this_monday}
\alias{this_monday}
\title{Helper to get the date for the Monday of the current week}
\usage{
this_monday()
}
\value{
Date for the Monday of the current week. For more details see \link[lubridate]{floor_date}.
}
\description{
Helper to get the date for the Monday of the current week
}
