% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_accuracy}
\alias{ts_accuracy}
\title{Compuate accuracy metrics in support of model selection}
\usage{
ts_accuracy(.data, horizon = 4, outcomes, .fun)
}
\arguments{
\item{.data}{Data to use for modeling}

\item{horizon}{Horizon of periods to use for splitting input to ".data" into training / test sets; default is \code{4}}

\item{outcomes}{Character vector specifying names of the column to use as the outcome}

\item{.fun}{List of modeling functions to use}
}
\value{
A `tibble`` containing one row for each combination of model and outcome with accuracy measures from \link[fabletools]{accuracy}.
}
\description{
Compuate accuracy metrics in support of model selection
}
