% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_cumulative_forecast}
\alias{ts_cumulative_forecast}
\title{Helper used in \code{ts_forecast()} to get cumulative foreacst from incident}
\usage{
ts_cumulative_forecast(.data, outcome = "cdeaths", inc_forecast)
}
\arguments{
\item{inc_forecast}{}
}
\description{
Helper used in \code{ts_forecast()} to get cumulative foreacst from incident
}
