% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_fit}
\alias{ts_fit}
\title{Wrapper to fit time series models}
\usage{
ts_fit(.data, outcomes, .fun, single = TRUE)
}
\arguments{
\item{.data}{Data to use for modeling}

\item{outcomes}{Character vector specifying names of the column to use as the outcome}

\item{.fun}{List of modeling functions to use}

\item{single}{Boolean indicating whether or not a "shortcut" should be used to return a single \code{tibble}; only works if there is one outcome ("outcomes") and one model function (".fun"); default is \code{TRUE}}
}
\value{
A single \code{mable} (model table) if (\code{single = TRUE}) or a named list of \code{mable}s (if \code{single = FALSE}). For more information see \link[fabletools]{mable}.
}
\description{
Wrapper to fit time series models
}
