% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_futurecases}
\alias{ts_futurecases}
\title{Helper to generate the estimate of incident cases from an icases forecast object}
\usage{
ts_futurecases(.data, .forecast, horizon = 4)
}
\arguments{
\item{.data}{Data from which the \link[tsibble]{new_data} should be generated; \emph{CAUTION} for best results make sure that the data passed to this argument is the same object as used to generate the model/forecast that is specified in ".forecast"}

\item{.forecast}{A \code{tibble} with forecast data generated using \link[focustools]{ts_forecast}; should \emph{only} be a forecast of incident cases}

\item{horizon}{Horizon periods through which the \link[tsibble]{new_data} should be generated; default is \code{4}}
}
\value{
A \code{tsibble} with horizon periods and respective forecasted incident cases.
}
\description{
This function takes a time series forecast and extracts the point estimate for incident cases out to a specified horizon. This is necessary to generate the "new_data" to be passed into the \link[focustools]{ts_forecast} incident death models that are based on lagged cases.
}
