% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission.R
\name{validate_forecast}
\alias{validate_forecast}
\title{Check for valid quantile csv file input}
\usage{
validate_forecast(filename, verbose = TRUE, install = FALSE, envname = NULL)
}
\arguments{
\item{filename}{Path to the forecast file to be checked}

\item{verbose}{Boolean indicating whether or not the output from this function should include validation message; default \code{TRUE}}

\item{install}{Boolean as to whether or not the python dependencies should be installed; if \code{TRUE} the module will be installed to the virtual environment specified in "envname"; default is \code{FALSE}}

\item{envname}{Character vector specifying the name of the virtualenv to which the python dependencies should be installed if \code{install = TRUE}; default is \code{NULL} which will install the module to a virtualenv named \code{r-reticulate}}
}
\value{
If \code{verbose = FALSE}, the returned value will be a boolean with \code{TRUE} for valid submission file and \code{FALSE} for invalid file. If \code{verbose = FALSE}, the function will return a named list with two elements: "valid" (boolean with the \code{TRUE}/\code{FALSE} validation code) and "message" (the output from the \verb{zoltpy valid_quantile_csv_file()} function).
}
\description{
Check for valid quantile csv file input
}
