% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{wis}
\alias{wis}
\title{Function to compute the Weighted Interval Score (WIS)}
\usage{
wis(forecast, truth, q = NULL)
}
\arguments{
\item{forecast}{A dataframe containing the model-predicted forecasts}

\item{truth}{A dataframe containing the observed values}

\item{q}{A vector of quantiles to assess the WIS over}
}
\value{
A dataframe containing the weighted interval scores
}
\description{
Function to compute the Weighted Interval Score (WIS)
}
