% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_arima_params}
\alias{extract_arima_params}
\title{Extract ARIMA parameters}
\usage{
extract_arima_params(arimafit)
}
\arguments{
\item{arimafit}{A single-row mable (`mdl_df`) from `focustools::model(arima=ARIMA(...))`.}
}
\value{
A single-row tibble containing ARIMA model parameter and diagnostics information.
}
\description{
Extracts ARIMA model parameters, including p, d, q, P, D, Q, and results from tidy() and glance() on an arima model object.
}
