% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forecast_pipeline}
\alias{forecast_pipeline}
\title{Pipeline to produce a forecast}
\usage{
forecast_pipeline(source = "nyt", granularity = "national", horizon = 4, ...)
}
\arguments{
\item{source}{data source (default for now: NYT)}

\item{granularity}{data granularity (default for now: national)}

\item{horizon}{Horizon periods through which the forecasts should be generated; default is \code{4}}

\item{...}{Arguments passed to other functions}
}
\description{
\lifecycle{experimental}

Runs the pipeline with reasonable defaults and some hard-coded values to do the following. See the Examples.
\enumerate{
\item Get data (national level from NYT by default)
\item Fit incident case and incident death models (ARIMA and lagged TSLM respectively)
\item Get future case data to create the incident death forecast
\item Create the incident death forecast based on this new data
\item Prepare submission format
\item Suggest a submission filename
\item Return all resulting objects to a list.
}
}
\examples{
\dontrun{
# Run all the steps to create models and forecasts
myforecast <- forecast_pipeline()
# Look at the submission and the suggested filename.
myforecast$submission
myforecast$submission_filename
# Write submission to file
readr::write_csv(myforecast$submission, file=myforecast$submission_filename)
# Validate submission
validate_forecast(myforecast$submission_filename)
}
}
