% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_cases}
\alias{get_cases}
\title{Retrieve case count data}
\usage{
get_cases(source = "jhu", granularity = "national")
}
\arguments{
\item{source}{Data source to query; must be one of \code{'jhu'} or \code{'nyt'}; default is \code{'jhu'}}

\item{granularity}{Data aggregation level; must be one of \code{'national'}, \code{'state'}, or \code{'county'}; if data source is \code{'nyt'} then only \code{'national'} can be used currently; default is \code{'national'}}
}
\value{
A \code{tibble} with (at minimum) the following columns:
\itemize{
\item \strong{epiyear}: Epidemiological year (see \link[lubridate]{epiyear} for more details)
\item \strong{epiweek}: Epidemiological week (see \link[lubridate]{epiweek} for more details)
\item \strong{icases}: Incident case count
\item \strong{ccases}: Cumulative case count
}

If \code{source = 'jhu'} and \code{granularity = 'state'} then the data will include column for \strong{state} (full name of the state) and \strong{date}. If \code{source = 'jhu'} and \code{granularity = 'county'} then the data will include column for \strong{fips} (county code) and \strong{date}.
}
\description{
Retrieve case count data
}
\references{
https://github.com/CSSEGISandData/COVID-19

https://github.com/nytimes/covid-19-data
}
