% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_cases}
\alias{get_cases}
\title{Retrieve case count data}
\usage{
get_cases(source = "jhu", granularity = "national", cache = TRUE)
}
\arguments{
\item{cache}{}
}
\description{
Retrieve case count data
}
\examples{

}
