% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Visualize and sanity check a forecast}
\usage{
plot_forecast(.data, submission, location = "US", pi = TRUE)
}
\arguments{
\item{.data}{Data used to create the submission}

\item{submission}{Formatted submission}

\item{location}{Vector specifying locations to filter to; "US" by default.}

\item{pi}{Logical as to whether or not the plot should include 50\% prediction interval; default is \code{TRUE}}
}
\description{

}
\examples{
\dontrun{
myforecast <- forecast_pipeline(force=TRUE)
plot_forecast(data=myforecast$data, submission=myforecast$submission, location="US")
}
}
