% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spread_value}
\alias{spread_value}
\title{Helper to reshape data for submission summary}
\usage{
spread_value(.data, ...)
}
\arguments{
\item{.data}{Tibble with submission data}

\item{...}{Additional arguments passed to \link[tidyr]{spread}}
}
\value{
Tibble
}
\description{
This function is used in \link[focustools]{submission_summary}. It spreads forecast targets to a wide format and forces "US" locations to be at the top of the resulting `tibble`.
}
