% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission.R
\name{submission_summary}
\alias{submission_summary}
\title{Summarize submission}
\usage{
submission_summary(.data, submission, location = NULL)
}
\arguments{
\item{.data}{Tibble with historical data for trend leading up to forecast}

\item{submission}{Formatted submission}

\item{location}{Vector specifying locations to filter to; \code{NULL} by default meaning all locations iwll be used}
}
\value{
List
}
\description{
Summarize submission
}
