% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_accuracy}
\alias{ts_accuracy}
\title{Compute accuracy metrics}
\usage{
ts_accuracy(.data, horizon = 4, outcomes, .fun)
}
\arguments{
\item{.data}{Data to use for modeling}

\item{horizon}{Horizon of periods to use for splitting input to ".data" into training / test sets; default is \code{4}}

\item{outcomes}{Character vector specifying names of the column to use as the outcome}

\item{.fun}{List of modeling functions to use}
}
\value{
A `tibble`` containing one row for each combination of model and outcome with accuracy measures from \link[fabletools]{accuracy}.
}
\description{
This function wraps \link[focustools]{ts_fit} to provide an interface for evaluating accuracy measures for one or more model definitions. Metrics included come from the \link[fabletools]{accuracy} function, and model definitions must be time series models (see \link[fabletools]{model} for more information). Note that the function uses the "horizon" to split data in training/test sets. For example, the default "horizon" (\code{4}) will use the last four weeks of input data as the test set and all data prior to those four weeks as the training set.
}
