% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_fit}
\alias{ts_fit}
\title{Wrapper to fit time series models}
\usage{
ts_fit(.data, outcomes, .fun, single = TRUE)
}
\arguments{
\item{single}{}
}
\value{
A \code{mable} (model table). For more information see \link[fabletools]{mable}.
}
\description{
Wrapper to fit time series models
}
