% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_forecast}
\alias{ts_forecast}
\title{Generate time series forecasts including quantile estimates}
\usage{
ts_forecast(mable, outcome, horizon = 4, new_data = NULL,
  seed = 1863, ...)
}
\arguments{
\item{mable}{A \code{mable} (model table); for more information see \link[fabletools]{mable}}

\item{outcome}{Name of the outcome; must be one of \code{'icases',},\code{'ideaths'}, \code{'cdeaths'}, \code{'ccases'}}

\item{horizon}{Optional horizon periods through which the forecasts should be generated; default is \code{4}}

\item{new_data}{Optional covariate data for forecasts using models that were fit using other variables; should be generated using \link[tsibble]{new_data}; default is \code{NULL}}

\item{seed}{Random seed used in bootstrapping process; default \code{1863}}

\item{...}{Additional parameters passed to the \link[focustools]{ts_cumulative_forecast} helper; only used if the forecast is cumulative}
}
\value{
A \code{tibble} with forecast results, including the name of the model, year and week, value of the forecast estimate, type of estimate (quantile or point), and bin of the quantile (if applicable) for the estimate.
}
\description{
Generate time series forecasts including quantile estimates
}
